Retire the grep-the-repo contract-test layer
- Delete Tests/RepoCommandContractTests.swift (4,445 lines): 103 suites that
  asserted string literals against 47 repo files, pinning scripts and docs to
  their exact source text. The genuinely valuable cross-artifact release
  checks (Info.plist vs appcast vs Homebrew cask, Sparkle feed settings, icon
  shipping, bundle-id compatibility) move to the new
  Tests/ReleaseMetadataContractTests.swift, which parses real values instead
  of grepping for phrases.
- Delete Tests/E2ESmokeSourceListContractTests.swift and
  Tests/FastTestCoverageContractTests.swift, which existed to grep
  check-build-source-lists.py and run-tests.sh.
- Gut scripts/dev/check-build-source-lists.py to its load-bearing check: the
  three hand-maintained shell source lists must point at files that exist.
  Its primary check compared a find-based list against an identical
  rglob-based list — tautological by construction — and its fast-test
  convention check duplicated what run-tests.sh already fails on before
  compiling. Name and exit contract unchanged, so CI and the test matrix
  keep working.
- Repoint the sample path in agent-check.py / test-matrix-checks.py
  self-tests at the new test file.
