+++ docs/posthog-product-intelligence-map.md
@@ -37,7 +37,7 @@ calls.
 | Can users trust speaker names? | `meeting_speaker_review_shown`, `meeting_speaker_review_submitted`, `meeting_speaker_match_reviewed`, `meeting_speaker_auto_recognized`, `meeting_speaker_finalization_failed`. | Later correction after an accepted auto-recognition is not isolated as a trust-decay metric. No content-level correctness can leave device. | Existing: `review_action`, `similarity_bucket`, `margin_bucket`, `call_count_bucket`, `channel`, `had_suggestion`, `review_item_bucket`, `known_people_bucket`, `completion_kind`, `updates_submitted_bucket`, `result`, `surface`. Add only `correction_window_bucket` if needed. | Add a dashboard query: override/correction rate by `similarity_bucket`, `margin_bucket`, and `had_suggestion`; weekly count of `meeting_speaker_finalization_failed`. | `SpeakerRecognitionTelemetry`, `SpeakerNamingSheet.swift`, `MeetingSessionController.swift` speaker finalization. | `Tests/SpeakerRecognitionTelemetryTests.swift` and `AnalyticsEventPolicyTests` speaker-review suites. | Small PR: add correction-window bucket if current reviewed/auto events cannot answer threshold tuning. |
 | Do retry and recovery paths actually recover users? | `workflow_recovery_attempted`, `workflow_recovery_finished`, `workflow_recovery_failed`, `workflow_abandoned`, dictation route recovery events, meeting failure/skipped events, `meeting_saved_audio_retranscription_requested`. | Wake/sleep and active-capture recovery are not fully represented in the product dashboard. Retry attempts are covered generically, but recovery-kind dashboards need stricter use. | Existing: `workflow_kind`, `failure_kind`, `retry_source`, `recovery_attempt_bucket`, `surface`, `artifact_retained`, `result`, `elapsed_bucket`, plus route/capture health buckets. Add `recovery_kind` only if `workflow_kind` is too broad. | `reliability.workflow_failure_rates`, `reliability.failure_kinds`, and a recovery outcome query grouped by `workflow_kind`, `failure_kind`, `retry_source`, `result`. Count totals from `workflow_recovery_finished`; use `workflow_recovery_failed` only for failure drill-downs. | `WorkflowRecoveryTelemetry`, `ProductFrictionTelemetry`, retry call sites in failed meeting and dictation flows. | `AnalyticsEventPolicyTests` pins recovery/friction fields; sanitizer tests cover raw error/path/url drops. | Medium PR: wire missing wake/active-capture recovery attempts if product wants recovery success rate, not just failure rate. |
 | Where does onboarding abandon? | `onboarding_shown`, `onboarding_step_viewed`, permission/model/CTA events, first dictation start/save/empty/stop, meeting dry run, agent CTA, reporting toggle, `onboarding_completed`, plus `workflow_abandoned`. | Mostly query shape. A single rollup event is optional; existing step events and abandoned workflow fields answer confident exits. | Existing: `step_id`, `step_index`, `flow_elapsed_bucket`, `step_elapsed_bucket`, `model_state`, `permission_kind`, `prior_status`, `to_status`, `cta`, `cta_type`, `first_dictation_saved`, `meeting_dry_run_completed`, `completion_flow`, `stage`, `reason_kind`. | `activation.ordered_funnel` plus an onboarding-friction query grouped by `step_id`, `stage`, `reason_kind`, and `flow_elapsed_bucket`. | `PermissionsOnboardingView.swift`; `ActivationTelemetry.trackWorkflowAbandoned` for confident exits. | `AnalyticsEventPolicyTests` pins onboarding funnel fields and sanitizer behavior. | Dashboard-only unless a specific abandoned-stage rollup proves materially easier to operate. |
-| Are users forming post-save habits? | `activation_second_artifact_saved`, `activation_return_proxy_observed`, `agent_capture_query_observed`, `dictation_completed`, `dictation_artifact_saved`, `meeting_transcript_saved`. | Habit quality is still aggregate/proxy. There is no client-emitted active-days bucket; the retention script computes habit cohorts from PostHog event history. | Existing: `days_since_first_bucket`, `first_artifact_kind`, `second_artifact_kind`, `return_window_bucket`, `prior_artifact_kind`, `artifact_kind`, `client_family`, `tool_kind`, `capture_kind`, `source_count_bucket`, `result_count_bucket`. Future: `repeat_modality`, `active_days_bucket` only if computed without IDs/content. | `scripts/ops/retention-cohort-report.py`, `wau.active_devices_by_week`, `wau.daily_active_devices`, `activation.return_windows`. | `ActivationTelemetry` saved-artifact/return proxy; retention scripts for aggregate cohorts. | `Tests/RepoCommandContractTests.swift` keeps first-value events in health probes/docs; retention script has self-test. | Dashboard/script-only for now. Medium PR only if PostHog-side cohorts cannot answer repeat modality. |
+| Are users forming post-save habits? | `activation_second_artifact_saved`, `activation_return_proxy_observed`, `agent_capture_query_observed`, `dictation_completed`, `dictation_artifact_saved`, `meeting_transcript_saved`. | Habit quality is still aggregate/proxy. There is no client-emitted active-days bucket; the retention script computes habit cohorts from PostHog event history. | Existing: `days_since_first_bucket`, `first_artifact_kind`, `second_artifact_kind`, `return_window_bucket`, `prior_artifact_kind`, `artifact_kind`, `client_family`, `tool_kind`, `capture_kind`, `source_count_bucket`, `result_count_bucket`. Future: `repeat_modality`, `active_days_bucket` only if computed without IDs/content. | `scripts/ops/retention-cohort-report.py`, `wau.active_devices_by_week`, `wau.daily_active_devices`, `activation.return_windows`. | `ActivationTelemetry` saved-artifact/return proxy; retention scripts for aggregate cohorts. | Retention script has self-test. | Dashboard/script-only for now. Medium PR only if PostHog-side cohorts cannot answer repeat modality. |
 | Are local summaries useful enough to keep? | `local_summary_requested`, `local_summary_finished`, `local_summary_failed`, `local_summary_cancelled`; legacy `local_meeting_summary_*` events may still appear during compatibility windows. Settings actions also cover retry/open/dismiss notice actions. | There is no direct feedback/usefulness event, and finished summaries are not yet part of the main product loop dashboard. | Existing: `provider`, `runtime`, `summary_action`, `setup_ready`, `queue_depth_bucket`, `chunk_count_bucket`, `duration_bucket`, `failure_kind`, `stage`. Add `result` or `feedback_bucket` only if a real user-visible feedback control exists. | Add a local-summary funnel: transcript saved -> summary requested -> finished/failed/cancelled -> opened/copied via settings actions. | `TranscriptedSettingsView.swift` summary request/finish/fail/cancel/action call sites. | `AnalyticsEventPolicyTests` pins local summary events through taxonomy parity; sanitizer tests cover forbidden payloads. | Small dashboard PR first. Product-code PR only when there is a real feedback action to instrument. |
 
 ## Dashboard Backlog
